Install React Native tools on vs code
This change was automatically added when I was
experimenting with using vsCode for development
insteal of sublime. Reason for this is due to the
fact that vsCode has better tooling for React (you
don't have to leave the tool to run commands, and
apparently, it has built in toold for debugging the
project), as well as being a better dev tool in
general.

Will experiment with it some more to figure out
the tool I'll continue with in the future.

(Other than insalling vsCode itself, all I added was
the reactNative tool and the support for javascript)