Installed react-navigation
I spent some time reading over the available
navigation tools. Breafly considered the two
that were recommended from the react
tutorial website
- native-navigation
- react-native-navigation

Was leaning on react-native-navigation
for a while due to its wide use but in the
end I decided to go for the actual react
tool thats still in the works called
react-navigation. It seems to be the
future of the system and will take over
in the near future. thought it best to ride
the hype while its still young